requirements.txt
biopython==1.76
decorator==4.4.2
joblib==0.14.1
networkx==2.4
numpy==1.18.3
pandas==1.0.3
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
